Support for more simple model loading: .ply file format

+++ sources/Core/EBO.cpp
@@ -24,6 +24,6 @@ GLuint EBO::generateNewEBO(std::vector<uint>* indices) {
     GLuint elementBufferObjectId = 0;
     glGenBuffers(1, &elementBufferObjectId);
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, elementBufferObjectId);
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, indices->size()*sizeof(GLfloat), indices->data(), GL_STATIC_DRAW);
+    glBufferData(GL_ELEMENT_ARRAY_BUFFER, indices->size()*sizeof(uint), indices->data(), GL_STATIC_DRAW);
     return elementBufferObjectId;
 }
